livemedia-creator: Reorganize the qemu arch patch
This should make it easier to support running lmc --virt-uefi on other
arches.

This also renames places using 'ovmf' a more correct 'fw_path' variable
name.

ovmf is the name of the x86_64 firmware, but other architectures use
different names, so rename this to --firmware and 'fw_path', but leaving
--ovmf-path as an alias.