lorax-composer: Catch dnf unknown group error
dnf-4.2.23-4 is raising a ValueError if the group doesn't exist, catch
that and add it to the error list when depsolving. Includes tests for
depsolve and freeze.

NOTE: Later versions of dnf (eg. at least dnf-4.4.2 and later) do not
raise any kind of error so these tests will fail on any system using the
newer dnf.

Resolves: rhbz#1943206