Install unicode.pf2 from new directory
grub2-2.06-3 changes where the unicode.pf2 font file is stored.
This changes the efi.tmpl to install it from the new location, which
means that it depends on grub2-2.06-3, but there is no way to express
this in the runtime-install.tmpl so if you see a failure like:

2021-07-08 16:10:05,586:   OSError: nothing matching /var/tmp/lorax/lorax.t80f74er/installroot/boot/grub2/fonts/unicode.pf2 in /

it means the new version of grub2 wasn't in the repos you used when
running lorax.

Fixes #1165