mkksiso: Rewrite to use xorriso features
xorriso is able to extract files from an iso, list files on the iso,
update the iso to change existing files and add new ones. It also has
the ability to preserve the bootloader images from the input iso. This
vastly simplifies the operation of mkksiso and removes the requirement
to run it as root.

(cherry picked from commit cf49a731b73e5f03da07af02552eca1198c7fc02)

Resolves: rhbz#2088631