mkksiso: Rewrite to use xorriso features
xorriso is able to extract files from an iso, list files on the iso,
update the iso to change existing files and add new ones. It also has
the ability to preserve the bootloader images from the input iso. This
vastly simplifies the operation of mkksiso and removes the requirement
to run it as root.