mkksiso: Fix grub2 editing error
Fix a dangerous bug in editing grub2 config files. A typo that left a
leading / in the path will lead to os.path.join not joining with the
tmpdir and editing the "/boot/grub2/grub.cfg" file on the host which
could lead to failure to boot when the system restarts, depending on
what was being added to the cmdline.