livemedia.ks: Add include for architecture specific packages
Add a %pre section to the livemedia.ks example that writes an include
file for architecture specific packages like the bootloader, or
workstation product, which is only available on x86_64.  This will allow
the example to be used on other arches as a text only live iso, with no
desktop.

Resolves: rhbz#2051548