livemedia-creator: Add --product to cmdline args
For reasons lost in the past lmc uses --project instead of --product
like lorax does to set the name of the iso. It also incorrectly says
that 'PROJECT' is substituted in the config files which has never been
true.

This fixes the documentation, and adds --product as an alias for
--project to make the cmdline args more consistent with lorax.