setup.py: Set the version number
Previously this was trying to import the version from pylorax. This
cannot work if the dependencies have not yet been installed, resulting
in a 'devel' version showing up in unexpected places. It actually never
really worked right.

Lorax uses tito for releases, and it understands how to update setup.py
version numbers, so drop the use of 'devel' and set the version to a
string.

Fixes #1278