libdnf5: Switch lorax to use libdnf5
At some point libdnf5 will be replacing libdnf. It is written in C++
with a SWIG interface, so accessing the API is quite different than
before.

This changes the template parser to use libdnf5 to install packages and
remove files. Tests pass, and building the boot.iso is working.

This simplifies the download callback a bit, we no longer have any idea what the total
download size is, so just print the package count and any errors.

The transaction callback logs install and script events.