rsyslogd.conf: Rewrite using current config syntax
The rsyslogd.conf file was written a long, long time ago. The format
that is was using is now called the the obsolete legacy format, so it
is time to update it.

It supports several levels of config commands, and they can mixed. So I
have converted the module loading and global config settings to
advanced, and left the templates and actions alone for now.

This also include setting the workDirectory, which fixes an error when
it tries to write the state file, and replaces the '~' action with
'stop'.