Pass vga=791 for live basic graphics mode on BIOS (#2176782)
As per https://bugzilla.redhat.com/show_bug.cgi?id=2176782 ,
this seems to be the best way to make Wayland work on basic
graphics mode on BIOS.

We don't need this for non-live, at least currently, as in the
traditional installer images, anaconda runs on X.org. We may
need to revisit this if we change the traditional installer env
to run on Wayland.

Signed-off-by: Adam Williamson <awilliam@redhat.com>