monitor: Skip repo errors involving CDROM/file source
When using the rhsm command anaconda first tries the cdrom, and when it
fails then tries the CDN. This causes lorax-composer to see the repo
failure and exit too early, preventing it from working with rhsm.

This should never show up during normal lmc usage, so it's safe to skip.
Includes new tests in LogMonitorTest