runtime-install: exclude renamed iwl firmware packages
@pbrobinson renamed the Intel wireless firmware packages in
the 20230625 build of linux-firmware. This causes installer
image builds to fail because they want to pull in the old packages
from the release repo because they fall in the "*-firmware" glob,
but those packages are now obsoleted by the newer ones from
updates/updates-testing, and dnf doesn't like that. This should
resolve the problem by specifically excluding all the renamed
packages from the glob. Of course, this change must go along
with the newer linux-firmware package, or else you'll get an
image that's missing a lot of wireless firmware.

Signed-off-by: Adam Williamson <awilliam@redhat.com>