ltmpl: Use ProcMount while running the dnf transaction
Some packages expect a filesystem that looks more like a real system. So
this mounts /proc and /dev inside the dnf installroot while packages are
being installed.

This was prompted by a systemd change which has been fixed
(https://github.com/systemd/systemd/pull/28388) but as noted, it may
happen again.

Also update tests to adjust for this. Sadly the run_pkg_transaction test
can now only be run as root. And when it doesn't run there is no /etc in
the installroot so make the directory just to be sure it is present.