runtime-install: install rsvg-pixbuf-loader
This was split out of librsvg2 as a subpackage recently, and
is only Recommended by the parent (not Required). For size's
sake, we don't pull Recommends into the installer env by
default. But we do really need this package, because quite a
lot of icons used in the installer are only present as SVGs in
the installer environment, so if we don't have the SVG loader we
can't show the icons.

Right now anaconda actually crashes if the SVG renderer isn't
present, which wasn't an expected outcome, but even if the crash
is fixed, the result would be that anaconda would run but not
show a lot of icons, which obviously isn't what we want. So we
should add this regardless of a fix for the crash.

Signed-off-by: Adam Williamson <awilliam@redhat.com>