ltmpl: Remove duplicate package objects from dnf5 results
In dnf5 if you have multiple repositories with overlapping content, and
the repo priorities are the same (eg. the default of 99) it will return
multiple results with the same nevra. The transaction will fail when
trying to install the duplicate rpms.

This de-duplicates the results from `_pkgver`, assuming that packages
with the same nevra and repo priority are interchangeable.

Note that this only works at an individual `installpkg` level, not across
the whole transaction. So it is possible for separate `installpkg`
commands that select the same package to cause a crash -- currently this
is not an issue, but is something to fix in the future.

Includes a duplicate package in the ltmpl test to make sure this stays
fixed.

Fixes #1356