templates.d/99-generic/live: Enable automatic persistence for live media
This change alters the live environment operating mode to default to
operating with full root system persistence using the functionality
recently added to Dracut.

It also adds menu entries so that it is possible to reset the persistence
or start without persistence at all.

This re-applies the change after it was reverted in Fedora Linux 38.

Reference: https://fedoraproject.org/wiki/Changes/ModernizeLiveMedia

Co-authored-by: Matt Coleman <matt@1eanda.com>