mkksiso: Add support for adding an anaconda updates.img
Anaconda supports an updates.img file that can be used to assist in
development, the files in it are added to the boot.iso's rootfs before
Anaconda is started.

Use `mkksiso --updates /path/to/updates.img` to add the image to the
iso. It can be used by itself or in combination with other options.

It would be great way to avoid requirement for HTTP server for local
development.