ltmpl: Handle installing provides with resolve_pkg_spec
Previously the installpkg command used filter_match to select packages,
which works great most of the time. It fails when the package is only
available as a provide. This happened recently with wget moving to
wget2-wget, causing builds to fail because it couldn't find wget.

This changes installpkg to use the resolve_pkg_spec function, which as
it happens also support globs and version comparison operators. DNF
doesn't like the `==` operator, but I have retained support for it by
translating it to a single `=`. I have dropped support for `=>` and `=<`
since they are never used and are and odd way to specify `>=` and `>=`
which of course still work.