dnf5: set releasever in the base object's vars
With Fedora 40 now branched, openQA installer build tests started
failing with an error:

  RuntimeError: cannot open file: (2) - No such file or directory [/etc/pki/rpm-gpg/RPM-GPG-KEY-fedora-$releasever-x86_64]

it seems like setting releasever via set_substitutions just does
not fully work as expected, and libdnf goes looking for a file
with a literal `$releasever` in it based on the gpgkey line in
the repo file. Though it seems like the releasever substitution
must work to *some* extent because we at least manage to resolve
and download packages before the failure.

This seems to fix it, anyhow, and is based on code from libdnf5
itself in dnf5/main.cpp RootCommand::set_argument_parser() .

Signed-off-by: Adam Williamson <awilliam@redhat.com>