dnfbase: Fix substitutions
Branching for Fedora 40 revealed a problem with using substitutions in
the GPG key name, @AdamWill fixed this by setting them on the Base
object which appears to be the correct way to do it.

Remove the per-repo kludge and set releasever and basearch on the Base
variables so they are used everywhere.