tito: Add RHEL tagger with support for rhbz and RHEL
This is a copy of the same tagger code from rhel9-branch but with
support for Jira RHEL-XXXX references added to it.

NOTE: The generated changelog will have leading '-' before the
Related/Resolves lines, this is added by tito after processing the
commits so they need to be removed manually while reviewing the
changelog.

Related: RHEL-31830