Replace squashfs_only with rootfs_type
The cmdline switch, --squashfs-only, now sets this and the new
--rootfs-type can be used to select other rootfs types. Currently only
squashfs and squashfs-ext4 are actually implemented.

This also changes the default to a plain squashfs without the wrapped
ext4 image. This speeds up image creation and there is no longer a
reason to use the wrapped ext4 method.

Related: RHEL-48606