livemedia: Add isomd5sum
This is needed for the iso checksum check of the live media. On the
boot.iso (and in Fedora) it is pulled in by anaconda-install-* packages,
but since liveinst isn't supported on RHEL we don't use those. Add it to
the kickstart to make sure the checksum service can run at boot time.

Resolves: RHEL-58896