erofs: Switch to using erofs by default
Also change the squashfs-tools Requires to Recommends so that it will
still be installed if it is available, and ignored if not.

Add checks so that --squashfs-only now conflicts with --rootfs-type when
using erofs. Also checks to make sure /usr/bin/mkfs.erofs is available,
or /usr/sbin/mksquashfs is available when parsing the cmdline so it will
exit early instead of running until rootfs creation fails without the
right tools.

Related: RHEL-48606