livemedia-creator: Set 0755 permission on / cpio overlay
livemedia-creator uses a cpio overlay on top of the inird to pass in the
kickstart when using virt to do the installation. When doing this the
default / permissions end up being 0700 can result in problems for
services running as non-root, eg. dbus, causing the install to hang.

Setting the permission of the overlay / directory to 0755 fixes this.

Thanks to jstodola for the debugging and fix of this.