mkksiso: Update to newer version supporting rebuilding efiboot.img
This rebases mkksiso to the same version as released in RHEL 10, it
includes the ability to rebuild the embedded efiboot.img which is needed
to set the kernel cmdline for USB booting. The does require being root,
and can be skipped when not needed by using '--skip-mkefiboot'

It also includes:
 * Replacing entries instead of just adding them
 * Adding an Anaconda updates.img and kernel cmdline entry
 * skipping UEFI boot image creation (useful for running as non-root on
   non-uefi systems).

The mkksiso tests have also been updated.

Resolves: RHEL-83208