runtime-cleanup: drop more video and audio firmwares
As noted in the comments above, we don't need to include video
and audio firmwares (in theory we might want audio for a11y
purposes, but sadly in practice we don't have any screen reading
implemented in the dedicated installer environment anyway).

dvb-firmware is for DVB broadcast receivers. cirrus and intel
-audio-firmware packages are, obviously, for audio devices.
intel-vsc-firmware is for the Intel Vision Sensing Controller,
which uses the optical sensor on a laptop not just as a webcam
but also proximity sensing and so on, which is all very well but
useless to an installer.

Signed-off-by: Adam Williamson <awilliam@redhat.com>