templates.d: Remove libdir variable from templates
Back in commit 84d67691a3fcb80a3fcf1d2cdbb1f05d930ce648 32 bit support
was removed and libdir set to 'lib64'.

There is no reason to pass the constant to the templates as a variable
anymore so this removes it from the code, stops passing it to templates,
and updates the templates to use 'lib64' where they were using the
variable.