Makefile: Stop running setup.py directly
Run pip build and pip install with no isolation so that they use the
system packages and don't try to connect to the network during the
build.

This now installs the templates to /usr/share/lorax as part of the
install target to compensate for pyproject's inability to install files
outside of the python package tree. Also install the config files
in /etc/ and /usr/lib/tmpfiles.d/