executils: Set encoding to UTF-8 for _run_program
This allows it to properly handle unicode output from the process that
is being run.

I have been able to reproduce the failure by running lorax in a VM with
a modified runtime-postinstall that calls a python program that outputs
unicode, or with glib2-2.85.3-1.fc43 -- but I am unable to cause it to
fail in the test suite.

(cherry picked from commit 3a5327868b201f2ab41fbf10e11b1d2633eb1b61)

Related: RHEL-121650