executils: Remove binary_output flag
This was never used in lorax, is always set to false, and now with
encoding set to UTF8 for these functions it wouldn't be binary anyway.

Also sets text=True in the _run_program call to startProgram, previously
the old universal_newlines was set from the state of binary_output.